requirements.txt
asgiref==3.8.1
Django==4.2.18
django-extensions==3.2.3
django-filter==23.1
djangorestframework==3.15.2
flake8==7.1.1
isort==6.0.0
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
